packages/analyzer/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python-headless==4.8.1.78
ffmpeg-python==0.2.0
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
Pillow==10.1.0
scenedetect[opencv]==0.6.2
httpx==0.25.2
openai-whisper>=20231117
whisperx==3.1.1
torch>=2.0.0
torchaudio>=2.0.0
faster-whisper>=0.9.0
# demucs==4.0.1  # Now using Docker image joanfont/demucs
